# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoSentinels — Environmental Intelligence API powered by Gemma 4
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface EcoHealthStatus {
  status: string;
  rag_ready: boolean;
  vector_store_docs: number;
  model: string;
  timestamp: string;
}

export type QueryRequestMode =
  (typeof QueryRequestMode)[keyof typeof QueryRequestMode];

export const QueryRequestMode = {
  rag: "rag",
  agent: "agent",
  alert: "alert",
} as const;

export interface QueryRequest {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  query: string;
  location?: string | null;
  lat?: number | null;
  lon?: number | null;
  mode?: QueryRequestMode;
  language?: string;
}

export interface Source {
  title: string;
  content: string;
  score: number;
  source_type: string;
}

export type RagResponseThreatLevel =
  | (typeof RagResponseThreatLevel)[keyof typeof RagResponseThreatLevel]
  | null;

export const RagResponseThreatLevel = {
  low: "low",
  moderate: "moderate",
  high: "high",
  critical: "critical",
} as const;

export interface RagResponse {
  answer: string;
  sources: Source[];
  threat_level?: RagResponseThreatLevel;
  recommendations: string[];
  agent_trace?: string[] | null;
  location?: string | null;
  timestamp: string;
}

export type SearchResponseResultsItem = {
  content: string;
  score: number;
  source_type: string;
  source_file: string;
};

export interface SearchResponse {
  query: string;
  results: SearchResponseResultsItem[];
}

export type AirQualityDataReadings = {
  [key: string]: {
    value?: number;
    unit?: string;
    last_updated?: string;
  };
};

export interface AirQualityData {
  station_name: string;
  distance_km: number;
  aqi?: number | null;
  aqi_category?: string | null;
  readings: AirQualityDataReadings;
}

export type WeatherDataForecast3dayItem = { [key: string]: unknown };

export interface WeatherData {
  temperature_c?: number | null;
  feels_like_c?: number | null;
  humidity_pct?: number | null;
  wind_speed_mph?: number | null;
  weather_description?: string | null;
  forecast_3day?: WeatherDataForecast3dayItem[];
}

export interface AnalysisRequest {
  location: string;
  lat: number;
  lon: number;
  include_air_quality?: boolean;
  include_weather?: boolean;
  detailed?: boolean;
}

export type AnalysisResponseCoordinates = {
  lat?: number;
  lon?: number;
};

export type AnalysisResponseOverallSeverity =
  (typeof AnalysisResponseOverallSeverity)[keyof typeof AnalysisResponseOverallSeverity];

export const AnalysisResponseOverallSeverity = {
  low: "low",
  moderate: "moderate",
  high: "high",
  critical: "critical",
} as const;

export interface AnalysisResponse {
  location: string;
  coordinates: AnalysisResponseCoordinates;
  air_quality?: unknown | null;
  weather?: unknown | null;
  risk_flags: string[];
  overall_severity: AnalysisResponseOverallSeverity;
  summary: string;
}

export type AlertRequestAlertType =
  (typeof AlertRequestAlertType)[keyof typeof AlertRequestAlertType];

export const AlertRequestAlertType = {
  air_quality: "air_quality",
  weather: "weather",
  wildfire: "wildfire",
  flood: "flood",
  heat_wave: "heat_wave",
  general: "general",
} as const;

export type AlertRequestSeverity =
  (typeof AlertRequestSeverity)[keyof typeof AlertRequestSeverity];

export const AlertRequestSeverity = {
  low: "low",
  moderate: "moderate",
  high: "high",
  critical: "critical",
} as const;

export type AlertRequestRawData = { [key: string]: unknown };

export interface AlertRequest {
  location: string;
  lat: number;
  lon: number;
  alert_type: AlertRequestAlertType;
  severity: AlertRequestSeverity;
  raw_data?: AlertRequestRawData;
}

export type AlertResponseThreatLevel =
  (typeof AlertResponseThreatLevel)[keyof typeof AlertResponseThreatLevel];

export const AlertResponseThreatLevel = {
  low: "low",
  moderate: "moderate",
  high: "high",
  critical: "critical",
} as const;

export interface AlertResponse {
  alert_id: string;
  title: string;
  summary: string;
  detailed_message: string;
  recommendations: string[];
  threat_level: AlertResponseThreatLevel;
  affected_populations: string[];
  timestamp: string;
}

export type RagSearchParams = {
  q: string;
  k?: number;
};

export type GetAirQualityParams = {
  lat: number;
  lon: number;
};

export type GetWeatherParams = {
  lat: number;
  lon: number;
};