# lib/api-zod/src/generated/types/airQualityData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoSentinels — Environmental Intelligence API powered by Gemma 4
 * OpenAPI spec version: 1.0.0
 */
import type { AirQualityDataReadings } from "./airQualityDataReadings";

export interface AirQualityData {
  station_name: string;
  distance_km: number;
  aqi?: number | null;
  aqi_category?: string | null;
  readings: AirQualityDataReadings;
}