# lib/api-zod/src/generated/types/alertRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoSentinels — Environmental Intelligence API powered by Gemma 4
 * OpenAPI spec version: 1.0.0
 */
import type { AlertRequestAlertType } from "./alertRequestAlertType";
import type { AlertRequestRawData } from "./alertRequestRawData";
import type { AlertRequestSeverity } from "./alertRequestSeverity";

export interface AlertRequest {
  location: string;
  lat: number;
  lon: number;
  alert_type: AlertRequestAlertType;
  severity: AlertRequestSeverity;
  raw_data?: AlertRequestRawData;
}