# lib/api-zod/src/generated/types/alertResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoSentinels — Environmental Intelligence API powered by Gemma 4
 * OpenAPI spec version: 1.0.0
 */
import type { AlertResponseThreatLevel } from "./alertResponseThreatLevel";

export interface AlertResponse {
  alert_id: string;
  title: string;
  summary: string;
  detailed_message: string;
  recommendations: string[];
  threat_level: AlertResponseThreatLevel;
  affected_populations: string[];
  timestamp: string;
}