# lib/api-zod/src/generated/types/analysisResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoSentinels — Environmental Intelligence API powered by Gemma 4
 * OpenAPI spec version: 1.0.0
 */
import type { AnalysisResponseCoordinates } from "./analysisResponseCoordinates";
import type { AnalysisResponseOverallSeverity } from "./analysisResponseOverallSeverity";

export interface AnalysisResponse {
  location: string;
  coordinates: AnalysisResponseCoordinates;
  air_quality?: unknown | null;
  weather?: unknown | null;
  risk_flags: string[];
  overall_severity: AnalysisResponseOverallSeverity;
  summary: string;
}