# lib/api-zod/src/generated/types/ecoHealthStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoSentinels — Environmental Intelligence API powered by Gemma 4
 * OpenAPI spec version: 1.0.0
 */

export interface EcoHealthStatus {
  status: string;
  rag_ready: boolean;
  vector_store_docs: number;
  model: string;
  timestamp: string;
}