# lib/api-zod/src/generated/types/ragResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoSentinels — Environmental Intelligence API powered by Gemma 4
 * OpenAPI spec version: 1.0.0
 */
import type { RagResponseThreatLevel } from "./ragResponseThreatLevel";
import type { Source } from "./source";

export interface RagResponse {
  answer: string;
  sources: Source[];
  threat_level?: RagResponseThreatLevel;
  recommendations: string[];
  agent_trace?: string[] | null;
  location?: string | null;
  timestamp: string;
}