# lib/api-zod/src/generated/types/weatherData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoSentinels — Environmental Intelligence API powered by Gemma 4
 * OpenAPI spec version: 1.0.0
 */
import type { WeatherDataForecast3dayItem } from "./weatherDataForecast3dayItem";

export interface WeatherData {
  temperature_c?: number | null;
  feels_like_c?: number | null;
  humidity_pct?: number | null;
  wind_speed_mph?: number | null;
  weather_description?: string | null;
  forecast_3day?: WeatherDataForecast3dayItem[];
}